requirements.txt
PySide2>=5.15.2.1,<6; platform_machine=="x86_64"
PySide6>=6.3,<6.4
qtpy>=2.1.0,<3
conan>=1.17.1,<2
graphviz>=0.19.1,<1
psutil>=5.9.0,<6
GitPython>=3.1.26,<4
qdarkstyle>=3.0.3,<4
validators>=0.18.2,<1
